+++ test/td_cache/rule_cache_test.exs
@@ -29,7 +29,7 @@ defmodule TdCache.RuleCacheTest do
   describe "RuleCache" do
     test "writes a rule entry in redis and reads it back", context do
       rule = context[:rule]
-      {:ok, [0, 1, "OK", 1]} = RuleCache.put(rule)
+      {:ok, [0, 1, 4, 1]} = RuleCache.put(rule)
       {:ok, s} = RuleCache.get(rule.id)
       assert s <~> rule
     end
@@ -37,7 +37,7 @@ defmodule TdCache.RuleCacheTest do
     test "adds a rule to the concept's rules", context do
       rule = context[:rule]
       concept = context[:concept]
-      {:ok, [0, 1, "OK", 1]} = RuleCache.put(rule)
+      {:ok, [0, 1, 4, 1]} = RuleCache.put(rule)
       {:ok, c} = ConceptCache.get(concept.id)
       assert c.rule_count == 1
     end
@@ -45,12 +45,12 @@ defmodule TdCache.RuleCacheTest do
     test "adds a rule to the concept's and then it is deleted", context do
       rule = context[:rule]
       concept = context[:concept]
-      {:ok, [0, 1, "OK", 1]} = RuleCache.put(rule)
+      {:ok, [0, 1, 4, 1]} = RuleCache.put(rule)
       {:ok, c} = ConceptCache.get(concept.id)
       assert c.rule_count == 1
       updated_at = DateTime.add(DateTime.utc_now(), 100)
 
-      {:ok, [1, "OK", 0]} =
+      {:ok, [1, 0, 0]} =
         rule
         |> Map.merge(%{business_concept_id: nil, updated_at: updated_at})
         |> RuleCache.put()